[mk] Improve `make update` strategy to better handle 'checkouts'
Change `make update` so that it updates the list of packages to be reinstalled
after each sucessful package. If a `make update` session fails in the middle of
a reinstallation, the next run will consider only packages that were not
handled correctly at the previous run.

On the one hand, there is no need to check at each run if a package was already
reinstalled, and this should handle 'checkouts' (whose version change each
second...) more cleanly. On the other hand, unclean 'work' directories left in
the middle of an update will probably raise a few more issues.