[mk] Add suport for INSTALL/DEINSTALL scripts
Those scripts are run by robotpkg_add/robotpkg_delete respectively, see their
manpage for details.

By default, a file named INSTALL or DEINSTALL in a package directory is added
to the list of scripts to be included in a package install/deinstall
framework. Additional scripts may be added by appending to INSTALL_SRC and
DEINSTALL_SRC. Scripts have @-style substitution. The content of INSTALL_VARS
is a list of variables whose value is to be substitued.