[mk] Fix a corner case situation regarding distfiles and checksums
When a distfile is already present (and has thus already passed the checksum
target in a previous run) but fails to pass the checksum target during the
current run, this means that either the distinfo file was updated in robotpkg,
or something bad happened to the file.

In any case, the boostrap-depends cookie has to be moved away becase one may
have to manually remove an obsolete/corrupted distfile and run fetch again with
updated boostrap dependencies (tnftp, wget, ...).

While here, update the 'checksum mismatched' error message to advise cleaning
up the faulty distfile, as the most probable scenario is that it needs to be
fetched again.