[mk/sysdep] Teach robotpkg about texinfo
This sysdep is twofold: it adds a build dependency on makeinfo et al. and a
full dependency on install-info.

It also provides a +INSTALL script framework to register/unregister info files
created by a package. The variable INFO_FILES must be set by a package to the
list of info files to be handled by the script (either relative to
${PKGINFODIR} or absolute paths).