[mk] Add a 'dir' format for checkouts
This extends MASTER_REPOSITORY or REPOSITORY.<pkg> to support a format of the
form "dir <path> <subdir>". This configures the checkout target to symlink
files in the directory "<path>/<subdir>" into "WRKSRC/<subdir>" (<subdir>
defaults to empty, i.e. ".").

Obviously, this does not make much sense to use this in a package global
MASTER_REPOSITORY. However, users may find it useful in a local
REPOSITORY.<pkg>. In particular, the 'checkout' regression test in
robotpkg/regress uses this.