[mk] Fix recovering from a failed update target in a 'checkout'ed package
Recovering from a previously failed update target in a 'checkout'ed package
implies recomputing the list of packages to update and starting all over
again. This is so because a 'checkout'ed package is always considered
outdated (there is no way to know if the package source was modified since the
last 'update' attempt).

This has no effect on the recovery of regular updates, because already
reinstalled packages will appear up-to-date (unless 'confirm' is given).