[sysutils/py-aafigure] Import aafigure-0.5
This package provides a python module ``aafigure``, that can be used from other
programs, and a command line tool ``aafigure``. ASCII art figures can be parsed
and output as SVG, PNG, JPEG, PDF and more.

For some reason, the package is not available for ubuntu-12.04, that's why it
is imported here.