[mk] Require update targets only when needed
update-vars.mk is required by the clean target, but this should not trigger the
inclusion of the rules for the 'update' target. Not only this is superfluous,
but since commit fc4521a this may also have the side-effect of creating a work
dir if the current package is outdated. Fix by requiring update rules from
robotpkg.mk only.