[mk] Detect moved/obsolete packages before starting a `make update`
While building the list of packages to update, abort if an installed package
cannot be found anymore in its PKGPATH. This happens for dependent packages of
the one being updated, and that have been moved or deleted. In this case, `make
update` won't be able to find the new location and will later fail in the
middle of the update. Better abort early and ask the user to manually
remove/update the problematic packages.