[image/artoolkitplus] Fix extraction on NetBSD with weird unzip
NetBSD-6 unzip does not support -b. By chance, the original unzip does not do
any conversion by default, so unzipping the files without -b at all works on
both version.

NetBSD unzip does not support -aa either (it interprets it as a redundant
-a). But -a on netbsd's unzip does not use the zip archive information and
tries to guess the nature of the file instead, so it also works by chance on
the artoolkit archive.

Sigh.