[mk] Relax strict requirement for a platform/${OPSYS}.mk file on Linux
When looking for platform/${OPSYS}.mk, fallback to platform/${OS_KERNEL}.mk
before raising a fatal error. This is mostly useful for Linux and the plethora
of available distributions that are all the same with just a different color on
the logo. It is still possible to define a more specific platform file
(e.g. Ubuntu.mk is kept).

In addition, drop platform files that were merely requiring a generic Linux.mk