[sysutils/rospack] Import rospack-2.0.13
rospack is a command-line tool for retrieving information about ROS packages
available on the filesystem. It implements a wide variety of commands ranging
from locating ROS packages in the filesystem, to listing available stacks, to
calculating the dependency tree of stacks. It is also used in the ROS build
system for calculating build information for packages.