[mk] Correctly handle multiple requirements for a dependency
It can happen that a dependency is required by multiple packages at the same
time. E.g. A depends on X and B, and B also depends on X. If A and B have
different requirements on X (different ABI or options requirements), they must
all be passed down to the build of X via PKGREQD.

When X is eventually built, those requirements must be merged as much as
possible into a single expression, so that the best matching flavour of X is
built (or conflicts detected).

This feature was not really useful until now because indirect dependencies are
usually not put in packages depend.mk files, so the scenario above does not
really happen with current packages. But handling this properly opens the path
to multi-packages, which will be required soon.