[image/openni-nite] Add INSTALL/DEINSTALL scripts for modules registration
Move the openi registration step into an INSTALL script, so that modules can be
unregistered when deleting the package - and so that binary package works.

Note: this leaves an additional dandling file in share/openi/licenses.xml, but
that file really belongs to middleware/openni and should be handled by an
INSTALL script in that package (as for modules.xml).

Bump PKGREVISION