[mk] Provide pre-bulk and post-bulk hooks
The pre-bulk (resp. post-bulk) hooks are called right before (resp. right
after) bulk bulding a package. pre-bulk is invoked before installing the
package dependencies, and post-bulk after the log summary has been generated in
${BULK_LOGIDR}/${PKGNAME}. When using a bulk set, hooks are invoked for each
_outdated_ package of a bulk-set (i.e. packages that are actually built), in
the context of the current package (i.e. PKGNAME et al. are set). They can
serve any purpose and are not used by robotpkg itself, thus they default to
noop.

The goal is to use this in the rbulk configuration file to provide a live
update of the rbulk database.