[mk] Suppress automatic recursion in the bulk target
Having the bulk target recursive raises a number of issues, especially when
already working with a package set. Bulk jobs with sets are normally already
sorted, so the need for a recursion within a single package is most probably
the sign of a missing dependency in the package Makefile. This must be
detected. Also, this might trigger unwanted builds outside the scope of a
'strict' set.

Since the automatic recursion can be trivially achieved by using the automatic
'-depends' set (e.g. make bulk-depends), there is absolutely no need for a
recursive bulk. So, say goodbye to ~20 lignes of obfuscated shell script.