[mk] Avoid returning an error when bulk is interrupted by a STOP file
Bulk interruption must exit with non-zero status so that when it is invoked as
part of a recursive set target, the parent make handling the set is also
interrupted. However, this does not distinguish between a clean interruption
and a fatal error, mostly because make(1) can only exit with 3 status codes
(0, 1 or 2) and does not propagate the sub-shell exit status.

So instead of raising a regular error, kill "make" itself (SIGINT makes sense
here, especially since it is cleanly caught by make). Finding the parent "make"
process from a Makefile rule is tricky, but ps has standard options for that,
so it's a least standard. The special exit status of "make" is then handled in
the recursive rule by printing a message and exiting with no error.