[mk] Fix error reporting when a set cannot be processed
The script tsort-sets.mk exited with status 2 in case of a fatal error while
processing a package Makefile. Since the script output is piped into other
processes, the exit code was discarded by the shell.

In addition to the exit code, the script must output the "***" string that is
understood by the reader process as an error.