[mk] Recheck depends when invoking a target with new depends in an unclean pkg
The depends cookie calls a callback(*) to check that no new dependencies
have been added since last resolution. This can happen if no package file
have changed but a different target is invoked (adding more dependencies)
in an existing WRKDIR.

This commit allows to avoid the ad-hoc removal of the boostrap-depends cookie
in checksum-file.

(*) because of ae4eb8d, this cannot be done directly in the depend.mk file