[image/openni-nite] Fix DISTFILES definition
This package has conditional DISTFILES, depending on the architecture.
The correct way to define this is to:
 - list all possible files in DISTFILES
 - let FETCH_ONLY and EXTRACT_ONLY point to the appropriate file, depending on
   the architecture.

This way, `make mdi` generates a distinfo for all files in one shot and `make
mirror-distfiles` could mirror everything - of course, this package has a
NO_PUBLIC_SRC, so mirror-distfiles will not mirror anything.