[mk] Let the bulk target do a mirror-distfiles
This is done in order to make use of the new 'check-master-site' feature, so
that bulk can report invalid MASTER_SITES or raise a warning whenever a remote
file has changed in one of the MASTER_SITES.

A new internal variable EXPECT_TARGETS in introduced to properly handle
conditional dependencies introduced by 'mirror-distfiles'. EXPECT_TARGETS let
robotpkg behave just as if the targets listed in the variable had been invoked
on the command line. In particular, this ensures that the list of dependencies
reported by print-depends is correct from the bulk point-of-view.