[mk] Locally cache the results of the 'preduce' macro
'preduce' is used, in particular, in alternative selection. The print-pkgnames
target thus makes indirects calls to this macro. If a package has a great
number of options, this can result in an overwhelming number of forks, because
preduce invokes an awk program.

For instance, robots/rflex-genom has 2304 different possible names and
`time make -C robots/rflex-genom print-pkgnames` is 27.68s user 11.58s system.
With this cache optimization, it becomes 2.02s user 0.04s system.