[mk] Fix progress count in a corner case situation in tsort-set.awk
In tsort-set.awk, an optimization when scanning each package consists in
adding a package list of dependencies right after having determined all the
package names, instead of making two separate calls to `make'. However, the
progress counter must be increased only if the optimized package was already
pushed on the todo stack.

This fixes the progress counter sometimes going beyond 100%.