[mk] Make the dependency on the fetch target optional in checksum and distinfo
The goal is to avoid the dependency on the fetch tool if all DISTFILES are
present. This used to be done in fetch.mk by checking the presence of
DISTFILES. However, this could sometimes be an issue because this made
the list of dependencies dependant on the context (typically, after a
successful fetch, the dependency on tnftp seems to vanish; this is problematic
for bulk).

So the fix is to always require the tnftp dependency in fetch.mk, but to
conditionally include the fetch target in 'checksum' or 'distinfo' (the only
two targets that require fetch). This way, the tnftp dependency is still
triggered only if needed but `make show-depends EXPECT_TARGETS='fetch'` always
show the correct list of dependencies.