[mk] Fix a potential infinite recursion in libdewey.awk
The expansion of '{}'-style patterns is processed recursively in libdewey.awk.
However, a potential infinite recursion could be triggered with a malformed
input consisting of single '{' brace (because a single opening brace does not
expand, or rather expands to itself). Fix this by detecting if an expansion
did not expand anything.

Note: This fixes the crazy awk symptom on Fedora, because Fedora is apparently
the only platform using brace-style patterns in mk/sysdep/py-setuptools.mk