[mk] Fix incorrect quoting of arguments in 'preduce' macro
If the dependency pattern to be reduced contained comas, they would incorrectly
be lost during a recursive $(eval ) call (because gmake would interpret them as
argument separators). Fix by correctly quoting the argument to $(eval).

Note: This is the second part of the 'crazy awk fix' on Fedora, that would
trigger the infinite recursion in 'preduce' by passing an invalid input
pattern.