[mk] Pass a list of expected targets for the bulk of a package set
The list of expected targets is required to correctly pick up all expected
dependencies of a target. Bulk will recursively run 'fetch', 'package' and
'mirror-distfiles'. This information has to be passed to the tsort-set script,
so that it can sort all dependencies before each package of a set runs its
'bulk' target and no new dependency is discovered lately (especially the
dependencies of the 'fetch' target).

Note: this introduces a somewhat private knowledge of the bulk target inside
the set-related targets, but no better fix has been found atm.