[hardware/serial] Fix 'configure' step relying on user's PATH
Do not use the bundled 'autogen' script from the sources, because it relies on
user's PATH to locate libtoolize, aclocal and autoconf.

Define a pre-configure target instead, invoking the required tools with their
full path, and then run the regular configure target.

Note: autoreconf cannot be used because the package don't use automake, and
that's automake who is in charge of install-sh (required by configure)...