[middleware/ros-actionlib] Fix build against system ros-base
There is an installation issue for generated .msg from .action, in
middleware/ros-genmsg. This is fixed in robotpkg. However, when compiling
actionlib against a system ros-base, this can only be fixed locally with a
post-install hook. The hook will be a noop when ros-base is compiled with
robotpkg.

This is temporary and it is fixed in groovy.