[robots/rmp-libs] Fix configure (for libtool>=2)
libtool>=2 requires that configure script define an m4 subdirectory, otherwise
libtoolize refuses to copy its scripts in the package. autoreconf is then
albe to produce a total mess without notice when the libtoolize version is
older than the version initialiy used for creating the distfile. Although this
is totally broken, this is how it (does not) works.

Add patch-ab and patch-ac fixing this. This fixes in particular the build of
robots/rmp-libs with the PREFER.libtool = robotpkg, as experienced by
ajgarcia@.