[supervision/openprs] Distribute version 1.1b8
A bunch of patches in the documentation (thanks to Piotr Trojanek) and one patch
in oprs-server-main (something with readline).
Resurect the truck-demo and make it autotools compliant.
Added the possibility to register "user" fd (on the top of the to select so we
wake up when needed).
Now use the XtAppAddInput mechanism to check for data comming on my sockets
Reduced the default_main_loop_pool from 100 Hz to 10 Hz.