[mk] Allow multiple instances of the same package in a package set
It is convenient to be able to specify multiple instances of the same package
(with different options) in a package set. For instance, it is interesting to
bulk build two instances of binary packages of a genom3 component using either
pocolibs or ros.

Instead of using only the first package match in tsort-set.awk, iterate over
the different patterns specified on the command line. Since tsort-set.awk is
already able to factor out things smartly, this seems to work flawlessly.

Note: commit c7a52d9 explains why only the first match was taken into account.
But currently the use case described there is not used nor required (afaik), so
we can ignore it. This commit should be reconsidered if there is actually a
need for such a use case.