[image/viam-dc1394-driver] Add a default version requirement in depend.mk
Spotted by David Gauchard, thanks!

Strictly speaking, the version requirement was not really missing, because when
depended on as a 'system' package, there is no way to retrieve the installed
version (the package has only one .so file). When depended on as a robotpkg
package, there is no problem though.