[graphics/assimp] Import assimp-3.0.1270
ASSIMP is a library to load and process geometric scenes from various
data formats. It is tailored at typical game scenarios by supporting a
node hierarchy, static or skinned meshes, materials, bone animations
and potential texture data. The library is *not* designed for speed,
it is primarily useful for importing assets from various sources once
and storing it in a engine-specific format for easy and fast
every-day-loading. ASSIMP is also able to apply various post
processing steps to the imported data such as conversion to indexed
meshes, calculation of normals or tangents/bitangents or conversion
from right-handed to left-handed coordinate systems.