[mk] Fix check-master-sites in case of a 3xx redirectrion and no content-length
It can happen that a MASTER_SITE first replies with a 3xx redirection followed
by a valid 2xx reply but with no content-length (http/1.1 allows this). In this
case, the content-length of the 3xx redirection must be properly ignored.

To do so, reset the variable holding the current content-length value when
encountering an http status line. This will not hurt for other kind of
transfers (ftp, file ...).