[mk] Update pkg_summary.gz during a `make package`
pkg_summary.gz is not used by any of the pkg_install tools, however it is
required by pkgsrc binary package management tools. Since it is trivially
generated, just do this during the `make package` target.

The only issue is that old packages will be retained in the pkg_summary.gz
file. But since old package .tgz files are also never pruned this is
consistent.

With this commit, pkgin works out of the box on a robotpkg repository.