[mk] Update the '-(o)-' log filter to let stderr pass through
Hiding stderr is a bad idea since this will typically hide compiler warnings or
other important errors. So let messages on stderr pass through the filter,
but continue logging them to the current log file (the relative order of
messages between stdout and stderr is not preserved, but this was also the case
before and this has never been a real issue in practice).

While here, drop the -r obfuscated option for logging the environment variables
(this will be reimplemented directly by robotpkg).