[mk] Invoke the '-(o)-' log filter automatically in relevant targets
Instead of implementing the log filtering in each individual rule, redefine the
SHELL variable used to invoke the rules recipes so that it points to the
appropriate log filter script, via a target specific variable.

The nice side effect is that package specific rules (pre-* or post-*
hooks or do-* targets) inherit from these target specific definitions and use
the log filter automatically.