[mk] Support '-c' option in the -(o)- log filter
gmake before 3.82 invokes $SHELL with the hardcoded '-c' option. It does not
honor the .SHELLFLAGS variable introduced in gmake-3.82.

Since the log filter is now used as $SHELL, it must thus support -c for compat
with gmake-3.81 (earlier gmake versions are not supported). Fortunately, the -c
options is not used in the log filter, so it can be easly implemented as an
alias for '--'.