[mk/robotpkgdb] Fix version computation after logfilter updates
A side effect of the logfilter used as the $SHELL command is that the $(shell)
gnumake function, when evaluated in the context of such a logged rule, also
uses $SHELL (this is legitimate).

Change the computation of this package version by precomputing it once for all
with immediate expansion. This is by the way also more efficient.