[mk] Allow 'make clean' in packages with a `make update` in progress
It can be useful to clean a failed build for a package in the middle of a `make
update`, without loosing the list of packages to be reinstalled.

make clean will now stop complaining about "update in progress" and simply
restore the "update in progress state" after cleaning the package. A warning is
echoed, with the list of packages not yet updated.

make clean confirm can be used to completely forget about the update in
progress state, as before.