[mk] Prevent potential recursion during `make update`
This is the final fix for the regression test introduced in commit a9c6a37 of
robotpkg/regress: a `make update` descending into an unclean package containing
a left over `update in progress` state may break badly because of a potential
recursion updating the original package.

Now, this is detected early and explicitly forbidden. An hopefully helpful
diagnostic message is printed. This involves passing down the list of PKGPATH
being worked on during dependency resolution, via the RECURSIVE_MAKE variable.