[metapkg/ros-base] Update to 1.9 (groovy)
Apart from the upgrade of individual packages, the only change is the
installation prefix. Now opt/ros/groovy is dropped and ros packages are
installed directly in LOCALBASE, since this release is FHS compliant
enough. The only change compared to debian ros packages is the location of
env.sh and setup.sh scripts that are installed in etc/ros instead of the root
prefix, for better FHS'ness. This has no consequence for ros itself (ros
does not use those files, they are user helper scripts only).

Also moved away actionlib and common_msgs packages from this meta-pkg, as
they're not part of ros-base (they belong to ros-desktop according to debian
packages).