[devel/boost-libs] Modularize boost-libs and enable component selection
Define a USE_BOOST_LIBS variable that may hold individual boost libraries to
check for. It defaults to a small selection of most useful(?) components,
including the ones that were previously searched for. It must be defined
before the last inclusion of boost-libs/depend.mk (or the default set is used).

For backward compatibility, the boost-libs/depend.mk behaviour is unchanged. In
particular, PREFIX.boost-libs and SYSTEM_FILES.boost-libs are still
available. But it should be noted that PREFIX.boost-libs can now contain more
than one value if the individual boost components are found in different
prefixes.