[mk] Fix an issue in 'deinstall' target not always deinstalling dependencies
In `make deinstall DEINSTALLDEPENDS=all`, the loop testing each dependency was
erroneously checking installed packages by using directly each tag of the
DEPEND_USE list. This is wrong, and DEPEND_ABI.<tag> must be used instead.

This fixes an issue with, in particular, all py-* and boost-lib-* packages
where the "tag" in DEPEND_USE and the actual package name are different.