[sysutils/console-bridge] Rename ros-console-bridge to console-bridge
This package was wrongly imported as ros-console-bridge; it should have been
console-bridge. Although it is provided by ros, this is a ros-independant
package... And there is actually a different ROS package called
rosconsole-bridge that do depend on ROS.

Beware, the distfile has also changed. The former one was the ros-ified version
of console-bridge, available on ros-gbp (that defeats the purpose of having a
ros-independant package) and the new one is the ros-free one, available on ros.