[meta-pkgs/ros-base] Provide a 'ros' alternative, switching beteen ros versions
Provide a 'ros' alternative for packages depending on ros packages. The
alternative currently recognizes the 'fuerte', 'groovy' and 'hydro' ros
versions. This is mostly interesting on ubuntu distributions supported by ros,
with a PREFER.ros = system setting, since configuring ros with PREFER.ros =
robotpkg will enable only the 'groovy' version (this is the only version
provided by robotpkg). Also, not all alternatives may be available for a
given package, depending on its requirements.

The default preferences is "groovy fuerte", in that order.
As usual, see `make show-options` for details.