[pkgtools/pkgin] Import pkgin-0.6.4
pkgin is aimed at being an apt / yum like tool for managing pkgsrc (and
robotpkg) binary packages. It relies on robotpkg_summary(5) for installation,
removal and upgrade of packages and associated dependencies, using a remote
repository.

The package has a number of patches. patch-a* are generic and have been
submitted upstream for integration. patch-b* are robotpkg-specific, and deal
with the pkg_* -> robotpkg_* tool renaming, as well as ~options suffix
handling.