[localization] Import ros-robot-state-publisher-1.9.9
This ROS package allows to publish the state of a robot to tf. Once the
state gets published, it is available to all components in the system
that also use tf. The package takes the joint angles of the robot, as
well as a kinematic description (URDF file), as input and publishes the
3D poses of the robot links, using a kinematic tree model of the robot.
The package can both be used as a library and as a ROS node.