[devel/ros-filters] Import ros-filters-1.7.2
This library provides a standardized interface for processing data as a
sequence of filters. This package contains a base class upon which to
build specific implementations as well as an interface which dynamically
loads filters based on runtime parameters.

Note that this version is *not* the version released with ROS Groovy
(which was 1.6.0). The only diff between 1.6.0 and 1.7.2 being the
catkinization of the package, we include this version here.