[mk] Fix 'checkout's revision number so that it fits within a 32 bits integer
Since pkg_install>=20110805.6, the local PKGREVISION is taken into account as
it should. However, checkout revision number did not fit within a 32bits
integer and this made the version checks fail badly, as reported by Aurélie
Clodic (thanks!).

So discard the year from the checkout revision, so that it's only mmddHHMMSS.
This could be an issue only around Dec 31 at midnight, when robotpkg is on
vacation anyway. And even then it should work since downgrading a package is
not an issue either.

While here, move away the checkout revision from the PKGREVISION number and use
the "pl" (patch level) extention instead.