[middleware/omniORBpy] Fix build against system omniORB package
The debian package for omniORBpy does not contain omniORBpy.h header. So check
instead for omniORB/__init.py.

Also deal with the same omniidl_be/__init__.py installed by both omniORB and
omniORBpy (thanks to a comment by sdb in pkgsrc/net/omniORBpy). Use a
DYNAMIC_PLIST_DIRS in that directory to automatically include the __init__.py
if it's not registered by omniORB (e.g. because omniORB is configured as a
system package).

Regen a sorted PLIST and bump PKGREVISION.