[mk/sysdep] Add a PLIST_FILTER that adds compiled python files to the PLIST
For packages that do install compiled python files, create a PLIST_FILTER that
automatically adds .py{c,o} files to the PLIST, according to what will be
compiled (defined by the PYTHON_PLIST_COMPILE_PATTERN variable).

Consequently, .pyc,o files need not to be added to the PLIST, simplifying the
generation of PLISTs for python (no need to iterate twice on make print-PLIST
anymore). Old PLISTs will still work, though.

Note: this was required only for properly handling the PLIST of
middleware/omniORBpy, that cannot be static because omniORBpy installs some
python files depending on the presence of middleware/omniORB in the same
prefix.