[mk] Improve error message when an installed robotpkg pkg is missing some files
Do not filter the error message displayed when a robotpkg package is installed
but the SYSTEM_SEARCH.pkg can't find some files.

Sometimes, the problem is more complex than just a missing file (for instance,
it can be a mismatched version in a bad .pc file), and printing the complete
log can help a lot in finding the issue.