[mk] Make sure to remove duplicates in the final PLIST of packages
Since f23ea3e, the PLIST_FILTER, if any, is applied after the awk functions
removing duplicates. It is necessary to run an extra round of uniquification
after PLIST_FILTER.

This fixes in particular the deletion of old packages with static PLISTs
containing compiled python files (that are now duplicated by the dynamic python
hooks).