[mk] Configure TMPDIR while building debian binary packages
Set TMPDIR to WRKDIR when invoking pkgrepo2deb, as /tmp might not be the best
place to store temporary files. First, the package may be big and /tmp not big
enough. Second, robotpkg is supposed to write files only in WKRDIR.

[and for that matter, some bulk build machines are configured to have WRKDIR
in a 'tmpfs' filesystem, so it may be more efficient to write there]