[interfaces/jrl-mal] Update blas and lapack variables passed to cmake
It seems that passing BLAS_LIBRARIES and LAPACK_LIBRARIES to cmake (instead of
BLAS_blas_LIBRARY and LAPACK_lapack_LIBRARY) also works for overriding cmake's
idea of blas and lacack.

In addition, this fixes some issues on Gentoo and/or with cmake-2.8.6 which
have a broken FindBLAS.cmake and FindLAPACK.cmake, respectively. See commit
9b1a367 for details.