[image/opencv] Fix depend.mk for PREFER.opencv = system
Commit f67e0cc was wrong: old opencv headers are still installed by modern
opencv. This is just the debian package that is split in a weird manner and
both libopencv-dev and libcv-dev must be installed. Both are needed since most
of the packages using opencv never bothered with updating their includes path
to opencv2.

So re-add include/opencv in the SYSTEM_SEARCH, and while here mention the
SYSTEM_PKG names for supported platforms.