[devel/boost-libs] Enhance BOOST_LIB_SUFFIX computation
Support boost library version tags in the computation of BOOST_LIB_SUFFIX.
While here, also update the search order so that the most specific tags are
found first (i.e. -mt-<version>.so.<version> down to a bare .so).

Also add a consistency check ensuring that all libraries have the same
suffixes. This is not strictly required, though, and may create artificial
issues, so this might be amended in the future.