[image/viam-genom] Fix CPPFLAGS/LDFLAGS for jpeg
Don't set CPPFLAGS/LDFLAGS for jpeg directly, but use INCLUDE_DIRS.jpeg and
LIBRARY_DIRS.jpeg instead. This has the same effect, except that this filters
out standard system libraries which should not be added to CPPFLAGS/LDFLAGS.

This fixes the build of image/viam-genom against a system opencv provided by
a system ros-fuerte when the standard opencv package from the distribution is
also installed.