[mk/sysdep] Fix openjdk.mk for Gentoo
Portage installs its openjdk under /usr/lib/jvm/icedtea*, so accept that.
It's probably not worth making a distinction between true openjdk and icedtea,
although it would be easy to add an "alternative" for that.

Also update SYSTEM_PKG for Gentoo (virtual/jdk is a generic meta package, it
could be configured to install another jvm).