[graphics/ros-python-qt-bindings] Fix "sip helper" cmake script
Boost_LIBRARIES may contain "optmized" or "debug" keywords, but they work only
with the target_link_library() cmake function. Fix the build_sip_binding()
function so that it filters out these keywords before passing them on the
command line.

This fixes the build of this package on Gentoo. Not sure why this is the only
platform that behaves like so, though.

Bump PKGREVISION.