[pkgtools/pkg_install] Depend on pax only for the 'package' target
The 'package' target needs pax for building the bootstrap.tar.gz archive, but
pax is otherwise not needed. So depend on it just when building a binary
package.

This is especially useful when PREFER.pax = robotpkg is used, as this avoids to
rely on a special 'inplace' depend.mk just for that (rare) situation.