[mk] Fix separation logic between build.mk and build-vars.mk
Cleanup build-vars.mk so that it does not define any rule, let build.mk include
its required build-vars.mk (not the other way round). This is required to make
sure that targets that are not related to the build do not trigger (for
instance) any dependency resolution.

By design, *-vars.mk are supposed to define only variables, while the files
without -vars are supposed to enable rules (and related dependencies). For
historical reasons, this separation is not really achieved yet, but the build/
phase is now clean.