[mk] Fix slightly incorrect dependency extraction in the bulk-<set> target
The dependency extraction (and topological sort) in a bulk-<set> must take into
account the expected LOCALBASE setting (which is ${BULKBASE}) in order to
see the very same list of dependencies as the bulk target. To do so, update
the semantics of the '-t' argument of tsort-set.awk to mean "any argument to be
passed down to ${MAKE}" (instead of just the list of expected targets) and use
that to pass LOCALBASE.

At the moment, the only visible effect is to change the default value of
PREFER.pkg_install (which is automatically set to system). But this avoids
circular dependencies with any bootstrap dependencies of pkg_install
(currently, at most archivers/pax and pkgtools/pkgrepo2deb for the 'package'
target, but this could grow).