[pkgtools/pkg_install] Save a few useless 'builds' but add a missing one
It is necessary to use the newly built robotpkg_* tools only when upgrading
from an older version (or from the -0 bootstrap version, which is covered by
the same logic). Otherwise the already installed tools are just fine.

But if the new tools are needed, pre-bulk must be hooked to the build, since
'bulk' needs robotpkg_{add,delete}.