[mk] Fix print-PLIST to filter out duplicates after path cleaning and substs
It can happen that duplicate PLIST entries are not detected until after the
path are cleaned (e.g. a duplicated / in a path is removed).

This is a bit redundant with the sort -u process, but the entries should be
sorted before back substitution of variables, so the sorting cannot take place
at the very end.

Noticed by Arnaud Degroote, thanks!