[mk] Change the environment cleaning strategy
Replace the hack used for cleaning up the environment. The most painful
issue with the previous strategy was with the ?= operator not working.

In gmake, if a variable is defined in the environment, ?= does nothing
(although a plain = is effective, which is a bit counter intuitive).
But ?= assigments are sometimes necessary for allowing user settings from
robotpkg.conf to override defaults, and all the point of cleaning the env. is
to make sure that it won't affect robotpkg.

The new strategy consists in just restarting gmake from an (almost) empty
environment (env -i). This is cleaner and lets gmake behave exactly as if the
filtered variables had never been present. In particular, ?= works as
expected.

While here, fix a long standing issue with $HOME. HOME is not filtered anymore,
so that ${HOME} can be used normally in robotpkg.conf. THe HOME variable is
now simply reset in the ALL_ENV environment which is passed to the
fetch/configure/build/install/etc. sub processes.