[mk] Update 'mirror-distfiles' so that it behaves like 'bulk'
The rationale is to separate the checks for MASTER_SITES and the bulk builds
proper. First, it is not necessary to test MASTER_SITES for all supported
platforms, where bulk builds are run. Second, sporadic MASTER_SITES failures
should not pollute the build logs or the production of packages.

This commit introduces MIRROR_LOGDIR (defaults to LOCALBASE/var/log/mirror),
where the 'mirror-distfiles' target will log into bulk-compatible format (so
that rbulk-report can be used transparently). Also, a set-aware
mirror-distfiles-<set> is created, so that one can control which packages are
mirrored & checked. Finally, mirror-distfiles is removed from the list of bulk
targets.